Added response data for retrieving coin list data and data for each coin, refactored method for training models for listed coins, refactored the method for fetching new coin data, fixed the training operations

+++ src/api/config/schedule.py
@@ -1,12 +1,13 @@
 from base64 import b64encode
 from datetime import datetime
-from os import environ, path, walk
+from os import environ, makedirs, path, walk
+from time import sleep
 
 from apscheduler.schedulers.background import BackgroundScheduler
 from pandas import DataFrame, json_normalize, read_csv
 from pycoingecko import CoinGeckoAPI
 
-from definitions import DATA_EXTERNAL_PATH, ROOT_DIR, app_name_env
+from definitions import DATA_EXTERNAL_PATH, ROOT_DIR, app_name_env, coins
 from modeling import train_coin_models
 from preparation import trim_dataframe
 from .git import append_commit_files, merge_csv_files, update_git_files
@@ -40,27 +41,37 @@ def data_dump():
 
 @scheduler.scheduled_job(trigger='cron', day=1)
 def model_training():
-    coin_list = read_csv(path.join(DATA_EXTERNAL_PATH, 'coin_list.csv'))
+    coin_list = read_csv(path.join(DATA_EXTERNAL_PATH, 'coin_list.csv')).to_dict('records')
     for coin in coin_list:
-        train_coin_models(coin)
+        if coin['id'] in coins:
+            train_coin_models(coin)
 
 
 @scheduler.scheduled_job(trigger='cron', hour=0)
-def update_coin_info():
+def fetch_coin_info():
     coin_gecko = CoinGeckoAPI()
     coin_list = coin_gecko.get_coins_list()
     json_normalize(coin_list).to_csv(path.join(DATA_EXTERNAL_PATH, 'coin_list.csv'), index=False)
     for coin in coin_list:
+        if coin['id'] not in coins:
+            continue
+
         updated_coin_data = coin_gecko.get_coin_market_chart_by_id(coin['id'], 'usd', 1)
-        updated_coin_dataframe = DataFrame(updated_coin_data, columns=['time', 'value'])
+        updated_coin_dataframe = DataFrame(updated_coin_data['prices'], columns=['time', 'value'])
         trim_dataframe(updated_coin_dataframe, 'time')
-        coin_dataframe = read_csv(path.join(DATA_EXTERNAL_PATH, coin['symbol'], 'data.csv'))
-        last_timestamp = coin_dataframe['time'].iloc[-1]
-        updated_coin_dataframe.drop(
-            index=updated_coin_dataframe.loc[updated_coin_dataframe['time'] < last_timestamp].index, inplace=True,
-            errors='ignore')
-        coin_dataframe.append(updated_coin_dataframe, ignore_index=True).to_csv(
-            path.join(DATA_EXTERNAL_PATH, coin['symbol'], 'data.csv'), index=False)
+        if path.exists(path.join(DATA_EXTERNAL_PATH, coin['symbol'])):
+            coin_dataframe = read_csv(path.join(DATA_EXTERNAL_PATH, coin['symbol'], 'data.csv'))
+            last_timestamp = coin_dataframe['time'].iloc[-1]
+            updated_coin_dataframe.drop(
+                index=updated_coin_dataframe.loc[updated_coin_dataframe['time'] < last_timestamp].index, inplace=True,
+                errors='ignore')
+            coin_dataframe.append(updated_coin_dataframe, ignore_index=True).to_csv(
+                path.join(DATA_EXTERNAL_PATH, coin['symbol'], 'data.csv'), index=False)
+        else:
+            makedirs(path.join(DATA_EXTERNAL_PATH, coin['symbol']))
+            updated_coin_dataframe.to_csv(path.join(DATA_EXTERNAL_PATH, coin['symbol'], 'data.csv'), index=False)
+
+        sleep(1)
 
 
 def schedule_jobs():
